Move TotalRAM() to Platform/Hardware.h with platform specific implementations.
